Add support for multiple scenes with central scene
When using central scene, the scene label and index specify what scene is
triggered and the value specify what the key is used. For example if a button
on a remote is pushed multiple times.

Previously the scene_id was set to the value index, and the scene_label was set
to the string value of the value.
Now the keys are set as following:
  - scene_id: the scene index (integer)
  - scene_label: the scene label (string)
  - scene_value_id: the integer value from the scene value (integer)
  - scene_value_label: the string value from the scene value (string)